feat(cli): add cli to honeybee
This commit adds a minimal but extendable command line interface to honeybee.

I'm using click to create the `cli` which gives us a great level of flexibility and fits our needs for ladybug tools.

Since the command line interface might not be needed for all the installations I added click to setup.py as an `extra_requires` for `cli`. One needs to use `pip install honeybee-core[cli]` or `pip install honeybee-core[all]` to install both honeybee core and the cli dependencies.