feat(logger): add logger
Added a logger that should be used from now on instead of using print.

After I started adding CLI it was obvious that all these random prints can end up being very annoying and handling stdout messages is very hard. This change will put all of that under control!

One question is if we should move the logging module to ladybug_core instead so we can use the same format / structure everywhere. For now I'm keeping it here and will use it for honeybee-radiance.