docs(properties): Add documentation to hidden properties methods
I hope it's ok that I also turned coveralls on for this repo with this PR. I know it can be annoying and we can turn it off at some point if it irritates people.  I just want to get a sense of where we currently stand with coverage and which lines of code are not tested.