feat(room): Add property for Room multiplier
I know that multipliers clearly have a use in energy simulation but, after thinking about it, they are just as helpful in computing daylight metrics over buildings and so they really should live in honeybee-core. So this commit adds them as a property on the room.

I also took the opportunity to revise several of the docstrings to make it clear which core properties are set-able or not (following a convention that I originally saw in RhinoCommon). You don't really have to review these though, @mostapharoudsari . Just looking over the Room changes and letting me know if you agree is good.