feat(adjacency): Output adjacency information from solve_adjacency
After deliberating a bit on different ways to do this, it seems best to just return the information about the adjacencies that were found during the solve_adjacency run. This way, the information can be used to assign custom properties to the newly-adjacent faces, apertures and doors. For example, assigning custom interior materials/constructions or setting all new adjacent faces to be air walls.