feat(add_prefix): Include option to add prefix to extension attributes
The recent shift in the setup of HVAC systems has created a new situation that I am sure will arise again in the future. Specifically, when a user wants to change the name of an object and all of its sub-objects (like after duplicating and moving a Room to combine it with the original Room into one Model), there are also some extension attributes that should be renamed (like single-room HVAC systems). This commit adds in code to call such add_prefix methods on extension attributes if the method exists. If the method does not exist, the renaming process simply proceeds as usual.