feat(shade): Add property to track whether Shade is on the indoors
This is_indoor property is needed to correctly set up the ModifierSet on honeybee-radiance.

Also, as part of this commit, I am adding the ability to assign Shades to Doors. This has now come up enough times from more than one person that I've realized that I better implement it now while I still have a near-complete idea of all the other repositories affected by the change (another month or two and I don't think that I would be able to catch all of the places that will be incorrect as a result of this change). As far as I can tell, the following repos need to be changed with a PR that parallels this one:
* honeybee-energy (especially the Model objectand Model writer)
* honeybee-schema (including both Door schema and the Wiki)
* energy-model-measure (OpenStudio has no concept of assigning Shades to Doors but I'll figure something out)
* honeybee-grasshopper-core (most of the visualization components and  the component that assigns Shade to a parent)

Lastly, I added a new method to the Face to add repeating windows at a fixed width x height now that we have methods for this on ladybug_geometry. This doesn't require any parallel PRs in other repos but, at some point, I should probably add a WindowParameter object to Dragonfly that makes use of this new method.