feat(model): Add properties for units_system and tolerance
This commit allows people creating Model objects to store the units system and tolerance on the Model itself. The aim of this is that, by specifying these values here, users don't need to worry about tedious geometry checks when they create dictionaries they want to serialize into Models. We can also use this to automatically scale the model to meters before export to EnergyPlus.