feat(transform): Ensure extension properties are transformed with host
This commit adds methods that allow extension attributes that are geometric to be transformed whenever their host is transformed.  These methods work by searching through the extension attributes and, if they find an extension that has the given transform method implemented (eg, move, rotate, scale), it will perform that transform whenever the host is transformed.

This is a necessary precursor to adding attributes to honeybee-radiance that include geometry (like dynamic shades that can be applied to specific radiance states). So, now, if someone moves a Room, the dynamic shades in the radiance properties of the Room's apertures are also moved.

@mostapharoudsari will also be very happy to see that the spell-checker is being put to good use.