ci(dependabot): Add config file for Dependabot
Adding this config file means that, whenever there is a new release of ladybug-core on PyPI, dependabot will immediately send a PR to this repo to ensure that there is a release of honeybee-core that is compatible with this version of ladybug-core.

Given that we have auto-merging turned on for this repo, this PR should be automatically merged if the tests pass, meaning there will automatically be a new release of honeybee-core whenever we release ladybug-core.