fix(face): Fix sub-face normal correction upon re-serialization
The way the Face.from_dict() method was originally written, the checks to ensure that child faces have the same normal as their parent face were not occurring.

This makes sure that this check happens and it still does the check to make sure that Ground or Adiabatic BCs aren't assigned to Faces with Apertures or Doors.