feat(model): Add properties and methods for managing glazing ratio
This commit adds several properties to the Model to make it easier to get the model-wide glazing ratio and to set glazing ratios for vertical fenestration and skylights.

It also adds more validations commands to separately validate the geometry and basic properties of the model.

And, while I realized that no other library was using these get_object_by_identifier methods, I deiced to drop the get_ while I still can.