feat(cli): Add more commands to generate different types of shades
This commit adds commands to add extruded borders and various types of louvers.

I also added a max_count option to the louvers by distance methods, which I think a lot of people will appreciate in order to generate light-shelf-like louvers.

LAstly, I'm changing the default WWR for the shoe box to be 0 so that it can better integrate with the edit methods that assign specific window and shade geometries.