fix(model): Handle the case of missing file extension
It seems that the luigi process under the hood does not use file extensions. So we will need a careful way to sense the file type without loading the whole file into memory. From what I can tell, sensing the { at the start of a JSON should be a safe bet.