fix(room): Ensure faces_by_guide_surface works with concave faces
I realized that the method might not be so reliable for concave faces where the center point lies outside of the face boundary.

So I am changing it to use the exact ladybug_geometry method that gets called in the solve adjacency method, which ensure that we always get a point that is on the face.