feat(validate): Greatly improve validation command
Now that I have a better sense of how we want to report validation issues, I have refactored all of the checking functions to return strings of error reports rather than a boolean. This means that users get a full report of all the issues found in the model rather than the first exception that was encountered.

I also added a method on the Model class that I imagine I will use in IronPython and is also used by the validation command. This method (and the command) can also be used to check extension attributes.