fix(room): Add a hidden property for Room._parent
I'm going to make use of this in Dragonfly and I thought about having Dragonfly extend the honeybee Room class but I think it's just much simpler and more efficient this way. I might also make use of it in other cases where I need to keep track of a particular building that the honeybee Room is apart of.